#1339 - Modularization of Action Bar
- Factored guidelines and preferences out of the various pages and into reusable components
