+++ webanno-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/guidelines/GuidelinesDialogContent.html
