+++ webanno-ui-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/curation/page/CurationPage.html
@@ -95,9 +95,7 @@
   <wicket:extend>
     <div wicket:id=openDocumentsModal></div>
     <div wicket:id=remergeDocumentDialog></div>
-    <div wicket:id="preferencesDialog"></div>
     <div wicket:id="exportDialog"></div>
-    <div wicket:id="guidelinesDialog"></div>
     <div wicket:id="finishDocumentDialog"></div>
 
     <div class="actionbar">
@@ -114,9 +112,6 @@
                 <li><a wicket:id="showExportDialog"> <img
                     src="images/document_export.png" title="Export"  /><br />Export
                 </a></li>
-                <li><a wicket:id="showPreferencesDialog"> <img src="images/cog.png"
-                    title="Settings" /><br />Settings
-                </a></li>
               </ul>
             </fieldset>
           </li>
@@ -128,15 +123,6 @@
                 <li><a wicket:id="toggleScriptDirection"> <img
                     src="images/text_width.png" title="Script direction" /><br />LTR/RTL
                 </a></li>
-              </ul>
-            </fieldset>
-          </li>
-
-          <li>
-            <fieldset>
-              <legend>Help</legend>
-              <ul class="horizontal">
-                <li wicket:id="showGuidelinesDialog" />
               </ul>
             </fieldset>
           </li>
@@ -186,9 +172,11 @@
                 </h3>
               </div>
               <div class="panel-body flex-v-container" style="padding: 0px;">
-                <nav class="flex-h-container">
+                <nav class="action-bar">
                   <span wicket:id="documentNavigator" />
                   <span wicket:id="pageNavigator"/>
+                  <span wicket:id="guidelinesDialog" />
+                  <span wicket:id="preferencesDialog" />
                 </nav>
                 <div wicket:id="mergeView" class="text-center flex-content flex-v-container"></div>
               </div>
